✨ Add Measurement From Binarized Image (#26)
* Added basic contouring

Added a function that performs basic image contouring and displays the contoured image. Also has some (currently commented) stuff to report the areas of all the contoured parts of the image, which will be used in the final IO.

* UI Refactor + Better Measuring

Refactored the UI to display the various levels of the image processing. In the next commit I will add display of measurements from the processed images.

* Added Measurement Reporting

That's all I am going to do on this branch. The next branch will be focused on UI Redesign so that auto-scaling components works the way it should. Right now it is admittedly not great. I will also likely branch for the directory/multi-image ui processing.